requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
openpyxl>=3.1.2
pandas>=2.0.3
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-filter>=23.2
xlrd>=2.0.1